Moved folders to prepare for github upload
