--- ILX_Lightwave_3724B.py
@@ -1,6 +1,0 @@
-import visa
-
-class ILX_Lightwave_3724B():
-    def __init__(self, visa_search_term):
-        rm = visa.ResourceManager()
-        self.inst = rm.open_resource(visa_search_term)
