+++ hardware/LICENSE.md
