+++ hardware/README.md
